Добавлена проверка на попытку удаления головной организации проекта в методе removeOrganizationFromProject. Теперь метод выбрасывает исключение, если удаляемая организация является головной, что предотвращает нежелательные изменения в структуре проекта и улучшает бизнес-логику управления проектами.

+++ app/Services/Project/ProjectService.php
@@ -517,6 +517,10 @@ class ProjectService
             throw new BusinessLogicException('Проект не найден в вашей организации.', 404);
         }
 
+        if ($organizationId === $project->organization_id) {
+            throw new BusinessLogicException('Нельзя удалить головную организацию проекта.', 400);
+        }
+
         $project->organizations()->detach($organizationId);
     }
